+++ build.sh
@@ -13,12 +13,18 @@
 # load the versions
 . ./loadenv.sh
 
-echo "Processing for JBOSS EAP Version $JBOSS_EAP"
+echo "Processing for JBOSS Data Grid Version $JBOSS_JDG"
+
+if [ ! -e install_files/jboss-datagrid-${JBOSS_JDG}-server.zip ]
+then
+    echo "could not find file install_files/jboss-datagrid-${JBOSS_JDG}-server.zip"
+    echo "You should put the required binary into the root directory first."
+    exit 255
+fi
 
 # Create containers
-echo "Creating JBoss EAP Container ..."
-docker pull jlgrock/centos-oraclejdk:6.6-8u45
-docker build -q --rm -t jlgrock/jboss.data-grid:$JBOSS_EAP .
+echo "Creating JBoss Data Grid Container ..."
+docker build -q --rm -t deloitte-irsm/rhjdg:$JBOSS_JDG .
 
 if [ $? -eq 0 ]; then
     echo "Container Built"